Beta 0.8(Dev-JA081024)
Fixed can't refresh token bug
